[Next Gen] refactor(codegen): split core patch-flag computation into patch_flag_inner
